requirements.txt
numpy>=1.19
pandas>=1.1
matplotlib>=3.3
scikit-learn>=0.23
transformers>=4.0
tqdm>=4.50
PyYAML>=5.3
requests>=2.24
